optimize

+++ bits.go
@@ -1,6 +1,6 @@
 package packet
 
-func makeMask(maskLength uint64) uint64 {
+func makeMask(maskLength uint) uint64 {
 	if 64 >= maskLength {
 		return 0xffffffffffffffff >> (64 - maskLength)
 	}
